@@ -12,21 +12,24 @@
   "private": true,
   "dependencies": {
     "@agm/core": "^1.0.0-beta.5",
-    "@angular/animations": "^6.0.9",
-    "@angular/cdk": "^6.3.3",
-    "@angular/common": "^6.0.3",
-    "@angular/compiler": "^6.0.3",
-    "@angular/core": "^6.0.3",
-    "@angular/forms": "^6.0.3",
-    "@angular/http": "^6.0.3",
-    "@angular/material": "^6.3.3",
-    "@angular/platform-browser": "^6.0.3",
-    "@angular/platform-browser-dynamic": "^6.0.3",
-    "@angular/router": "^6.0.3",
+    "@alyle/ui": "^2.6.0",
+    "@angular/animations": "7.2.1",
+    "@angular/cdk": "7.2.1",
+    "@angular/common": "7.2.1",
+    "@angular/compiler": "7.2.1",
+    "@angular/core": "7.2.1",
+    "@angular/forms": "7.2.1",
+    "@angular/http": "7.2.1",
+    "@angular/material": "7.2.1",
+    "@angular/platform-browser": "7.2.1",
+    "@angular/platform-browser-dynamic": "7.2.1",
+    "@angular/router": "7.2.1",
     "@ngrx/core": "^1.2.0",
     "@ngrx/store": "^6.0.1",
+    "angular": "^1.7.8",
     "angularfire2": "^5.0.0-rc.11",
     "chart.js": "^2.7.3",
+    "chroma-js": "^1.3.6",
     "core-js": "^2.5.4",
     "firebase": "^5.3.0",
     "hammerjs": "^2.0.8",
@@ -35,23 +38,24 @@
     "ngx-image-cropper": "^1.0.2",
     "ngx-img-cropper": "^0.10.4",
     "ngx-moment": "^3.2.0",
-    "ngx-spinner": "^6.1.2",
+    "ngx-spinner": "^7.0.0",
     "primeicons": "^1.0.0",
     "primeng": "^7.0.0-beta.1",
-    "rxjs": "^6.3.2",
+    "rxjs": "^6.0.0",
     "rxjs-compat": "^6.3.2",
+    "tsc": "^1.20150623.0",
     "zone.js": "^0.8.26"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "~0.6.8",
-    "@angular/cli": "~6.0.8",
-    "@angular/compiler-cli": "^6.0.3",
-    "@angular/language-service": "^6.0.3",
+    "@angular/cli": "7.2.1",
+    "@angular/compiler-cli": "7.2.1",
+    "@angular/language-service": "7.2.1",
     "@types/googlemaps": "^3.30.16",
     "@types/jasmine": "~2.8.6",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^8.9.5",
-    "codelyzer": "~4.2.1",
+    "codelyzer": "^4.5.0",
     "jasmine-core": "~2.99.1",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~1.7.1",
@@ -62,6 +66,6 @@
     "protractor": "~5.3.0",
     "ts-node": "~5.0.1",
     "tslint": "~5.9.1",
-    "typescript": "^2.7.2"
+    "typescript": "3.1.1"
   }
 }
