@@ -38,6 +38,8 @@ import com.mylook.mylook.entities.Closet;
 import com.mylook.mylook.entities.Favorite;
 import com.mylook.mylook.entities.Outfit;
 import com.mylook.mylook.info.ArticleInfoActivity;
+import com.mylook.mylook.session.Sesion;
+import com.mylook.mylook.utils.OutfitAdapter;
 
 import java.util.ArrayList;
 
@@ -50,11 +52,35 @@ public class CategoryTab extends Fragment {
     private Context context;
     private GridView outfitGrid;
     private Activity act;
+    private String dbUserId = Sesion.getInstance().getSessionUserId();
     private FloatingActionButton addOutfit;
     private ProgressBar mProgressBar;
+    private static CategoryTab instance = null;
+    private static boolean loaded = false;
+    private OutfitAdapter adapter;
 
     public CategoryTab() {
         // Required empty public constructor
+    }
+
+    public static CategoryTab getInstance() {
+        if (instance == null)
+            instance = new CategoryTab();
+        return instance;
+    }
+
+    /**
+     * Método para cuando haya habido algun cambio y haya que actualizar los objetos
+     */
+    public static void refreshStatus(){
+        if(instance!=null){
+            loaded = false;
+        }
+    }
+
+    @Override
+    public void onResume() {
+        super.onResume();
     }
 
     @Override
@@ -67,6 +93,7 @@ public class CategoryTab extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
         act = getActivity();
+
         return inflater.inflate(R.layout.tab_categories, container, false);
 
     }
@@ -74,12 +101,19 @@ public class CategoryTab extends Fragment {
     @Override
     public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
         mProgressBar = view.findViewById(R.id.mProgressBar);
-        mProgressBar.setVisibility(View.VISIBLE);
+
         outfitGrid = view.findViewById(R.id.grid_colecciones);
-        setGridview();
         addOutfit = view.findViewById(R.id.addOutfit);
+        if (!loaded) {
+            mProgressBar.setVisibility(View.VISIBLE);
+            outfits = new ArrayList<>();
+            adapter = new com.mylook.mylook.utils.OutfitAdapter(act, outfits);
+            setGridview();
+            getOutfits();
+        }
 
 
+        outfitGrid.setAdapter(adapter);
         addOutfit.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
@@ -95,12 +129,10 @@ public class CategoryTab extends Fragment {
         outfitGrid.setColumnWidth(imageWidth);
         outfitGrid.setHorizontalSpacing(8);
         outfitGrid.setNumColumns(2);
-        getOutfits();
         outfitGrid.setOnItemClickListener(new AdapterView.OnItemClickListener() {
             public void onItemClick(AdapterView<?> parent, View v,
                                     int position, long id) {
                 loadOutfit(parent, position);
-
             }
         });
     }
@@ -108,19 +140,19 @@ public class CategoryTab extends Fragment {
     private void loadOutfit(AdapterView<?> parent, int position) {
         mProgressBar.setVisibility(View.VISIBLE);
         final String outfitId = ((Outfit) parent.getAdapter().getItem(position)).getOutfitId();
-        dB.collection("closets").whereEqualTo("userID",user.getUid()).get()
+        dB.collection("closets").whereEqualTo("userID", dbUserId).get()
                 .addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
                     @Override
                     public void onComplete(@NonNull Task<QuerySnapshot> task) {
-                        if(task.isSuccessful()){
+                        if (task.isSuccessful()) {
                             final String id = task.getResult().getDocuments().get(0).getId();
                             dB.collection("closets").document(id).collection("outfits")
                                     .document(outfitId).get().addOnCompleteListener(new OnCompleteListener<DocumentSnapshot>() {
                                 @Override
                                 public void onComplete(@NonNull Task<DocumentSnapshot> task) {
                                     Outfit outfit = task.getResult().toObject(Outfit.class);
-                                    Intent intent = new Intent(getContext(),ViewOutfitActivity.class);
-                                    intent.putExtra("items",outfit.getItems());
+                                    Intent intent = new Intent(getContext(), ViewOutfitActivity.class);
+                                    intent.putExtra("items", outfit.getItems());
                                     intent.putExtra("name", outfit.getName());
                                     intent.putExtra("category", outfit.getCategory());
                                     intent.putExtra("id", task.getResult().getId());
@@ -147,7 +179,7 @@ public class CategoryTab extends Fragment {
 
     private void getOutfits() {
         dB.collection("closets")
-                .whereEqualTo("userID", user.getUid())
+                .whereEqualTo("userID", dbUserId)
                 .get()
                 .addOnCompleteListener(new OnCompleteListener<QuerySnapshot>() {
                     @Override
@@ -162,13 +194,13 @@ public class CategoryTab extends Fragment {
                                             public void onComplete(@NonNull Task<QuerySnapshot> task) {
                                                 if (task.isSuccessful()) {
                                                     ArrayList<String> arrayList = new ArrayList<>();
-                                                    outfits = new ArrayList<>();
                                                     for (QueryDocumentSnapshot documentSnapshot : task.getResult()) {
                                                         Outfit outfit = documentSnapshot.toObject(Outfit.class);
                                                         outfit.setOutfitId(documentSnapshot.getId());
                                                         outfits.add(outfit);
+                                                        adapter.notifyDataSetChanged();
+                                                        loaded = true;
                                                     }
-                                                    outfitGrid.setAdapter(new com.mylook.mylook.utils.OutfitAdapter(act, outfits));
                                                     mProgressBar.setVisibility(View.INVISIBLE);
                                                     return;
                                                 } else
@@ -192,7 +224,6 @@ public class CategoryTab extends Fragment {
         input.setHint((CharSequence) "Nombre");
 
 
-
         LinearLayout linearLayout = new LinearLayout(getContext());
 
         LinearLayout.LayoutParams layoutParams = new LinearLayout.LayoutParams(ViewGroup.LayoutParams.MATCH_PARENT, ViewGroup.LayoutParams.MATCH_PARENT);
@@ -211,7 +242,7 @@ public class CategoryTab extends Fragment {
                     @Override
                     public void onClick(DialogInterface paramDialogInterface, int paramInt) {
                         String newOutfitName = input.getText().toString();
-                        Intent intent = new Intent(getContext(), OutfitActivity.class);
+                        Intent intent = new Intent(getActivity(), OutfitActivity.class);
                         intent.putExtra("name", newOutfitName);
                         intent.putExtra("category", "normal");
                         startActivity(intent);
