@@ -1,16 +1,15 @@
 package com.mylook.mylook.closet;
 
-import android.app.MediaRouteButton;
+import android.app.ActionBar;
 import android.content.DialogInterface;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
-import android.support.constraint.ConstraintLayout;
-import android.support.constraint.Guideline;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.RecyclerView;
 import android.support.v7.widget.Toolbar;
 import android.util.DisplayMetrics;
+import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
@@ -96,6 +95,7 @@ public class ViewOutfitActivity extends AppCompatActivity {
         tb = findViewById(R.id.toolbar);
         tb.setTitle(collectionName);
         setSupportActionBar(tb);
+
         getSupportActionBar().setDisplayHomeAsUpEnabled(true);
     }
 
@@ -116,6 +116,12 @@ public class ViewOutfitActivity extends AppCompatActivity {
             intent.putExtra("category",category);
             intent.putExtra("id",outfitId);
             startActivity(intent);
+            finish();
+        } else if (id==R.id.home) {
+
+
+        } else{
+            onBackPressed();
             finish();
         }
 
