@@ -1,9 +1,11 @@
-import { Component } from '@angular/core';
+import { Component, OnDestroy } from '@angular/core';
 import { AuthService } from '../../services/auth.service';
 import { Router, Params } from '@angular/router';
 import {FormBuilder, FormGroup, Validators } from '@angular/forms';
 import {FormControl, FormGroupDirective, NgForm } from '@angular/forms';
 import {ErrorStateMatcher} from '@angular/material/core';
+import {UserService} from '../../services/user.service';
+import { Subscription } from 'rxjs';
 
 @Component({
   selector: 'app-login',
@@ -11,8 +13,8 @@ import {ErrorStateMatcher} from '@angular/material/core';
   styleUrls: ['login.component.scss']
 })
 
-export class LoginComponent {
-
+export class LoginComponent implements OnDestroy{
+  _subscription: Subscription;
   loginForm: FormGroup;
   errorMessage = '';
   isLoading = false;
@@ -25,11 +27,20 @@ export class LoginComponent {
   matcher = new MyErrorStateMatcher();
 
   constructor(
+    public userService: UserService,
     public authService: AuthService,
     private router: Router,
     private fb: FormBuilder
   ) {
     this.createForm();
+  }
+
+  ngOnDestroy(): void {
+    if(this._subscription){
+      console.log("destruyendo ***************************************************************", this._subscription);
+      this._subscription.unsubscribe();
+    }
+   // 
   }
 
   createForm() {
@@ -39,40 +50,70 @@ export class LoginComponent {
     });
   }
 
+  getUserStore(userUID) {
+    return new Promise((resolve) => {
+      console.log("la subs ", this._subscription)
+      this._subscription = this.userService.getUserInfo(userUID).subscribe(userA => {
+        console.log("que carajo pasa aca ", userA )
+        resolve (userA[0].storeName)
+      })
+      });
+  }
+
   tryFacebookLogin() {
     this.authService.doFacebookLogin()
     .then(res => {
-      this.router.navigate(['/home']);
+      this.getUserStore(res.user.uid).then( res => {
+        this._subscription.unsubscribe();
+        this.router.navigate(['Tiendas', res]);
+        }, err => {
+          console.log(err);
+          this.errorMessage = err;
+        });
     });
   }
 
   tryTwitterLogin() {
     this.authService.doTwitterLogin()
     .then(res => {
-      this.router.navigate(['/home']);
+      this.getUserStore(res.user.uid).then( res => {
+        this._subscription.unsubscribe();
+        this.router.navigate(['Tiendas', res]);
+        }, err => {
+          console.log(err);
+          this.errorMessage = err;
+        });
     });
   }
 
   tryGoogleLogin() {
     this.authService.doGoogleLogin()
     .then(res => {
-      this.router.navigate(['/home']);
+      this.getUserStore(res.user.uid).then( res => {
+        this._subscription.unsubscribe();
+        this.router.navigate(['Tiendas', res]);
+        }, err => {
+          console.log(err);
+          this.errorMessage = err;
+        });
     });
   }
 
   tryLogin() {
     this.isLoading = true;
     console.log(this.loginForm.value);
-    this.authService.doLogin(this.loginForm)
-    .then(res => {
-     this.isLoading = false;
-     this.router.navigate(['/home']);
+    this.authService.doLogin(this.loginForm).then( res => {
+      console.log(`tamadre`, res);
+      this.getUserStore(res.user.uid).then((nombreTienda) => {
+        this.isLoading = false;
+        this._subscription.unsubscribe();
+        this.router.navigate(['Tiendas', nombreTienda]);
+      });
     }, err => {
       this.isLoading = false;
       console.log(err);
       this.errorMessage = err;
     });
-    console.log(this.errorMessage);
   }
 }
 
