@@ -2,7 +2,7 @@ import { Injectable } from '@angular/core';
 import {CanActivate, ActivatedRouteSnapshot, Router} from '@angular/router';
 import { AngularFireAuth } from 'angularfire2/auth';
 import { UserService } from './user.service';
-
+import { DataService } from '../../service/dataService';
 
 @Injectable()
 export class AuthGuard implements CanActivate {
@@ -10,17 +10,35 @@ export class AuthGuard implements CanActivate {
   constructor(
     public afAuth: AngularFireAuth,
     public userService: UserService,
-    private router: Router
+    private router: Router,
+    public dataService: DataService
   ) {}
 
   canActivate(): Promise<boolean> {
     return new Promise((resolve, reject) => {
-      this.userService.getCurrentUser()
-      .then(user => {
-        this.router.navigate(['/home']);
-        return resolve(false);
-      }, err => {
-        return resolve(true);
+      this.dataService.getStoreInfo().then((user)=> {
+        console.log('firebaseUID:', user.firebaseUID);
+        console.log('this.router.url', this.router.url.toString())
+        if(user.firebaseUID !== '' && !this.router.url.toString().includes('Tiendas')) {
+          console.log("*********************************** esta logeado y con url -" , this.router.url)
+          this.router.navigate(['Tiendas', user.storeName]);
+          return resolve(false);
+        } else {
+          if( user.firebaseUID !== '' && this.router.url.toString().includes('Tiendas')){
+            console.log("*********************************** esta logeado y con url +" , this.router.url)
+            return resolve(true);
+          } else {
+            if( user.firebaseUID === '' && this.router.url.toString().includes('Tiendas')) {
+              console.log("*********************************** esta logeado y con url +" , this.router.url)
+              this.router.navigate(['Inicio']);
+              return resolve(false);
+            } else {
+              console.log("*********************************** no esta logueado y no va a tienda")
+              return resolve(true);
+            }
+            
+          }
+        }
       });
     });
   }
