@@ -1,3 +1,11 @@
 export const environment = {
-  production: true
+  production: true,
+  firebase: {
+    apiKey: 'AIzaSyC7gRAUi_0C-w5gqsWIVp0IKyEZT-3UiUA',
+    authDomain: 'mylook-develop.firebaseapp.com',
+    databaseURL: 'https://mylook-develop.firebaseio.com',
+    projectId: 'mylook-develop',
+    storageBucket: 'mylook-develop.appspot.com',
+    messagingSenderId: '259536885025'
+  },
 };
